# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: TriReservas API
  description: |-
    TriReservas es una aplicación web para la gestión de alojamientos y reservas, con roles diferenciados para 
    usuarios y anfitriones. Desarrollada con Spring Boot, JPA y Angular, ofrece una solución moderna y eficiente para la
    administración de estancias.
    
    Requisitos Funcionales:
    * [Aquí](https://docs.google.com/spreadsheets/d/14pBo766cm17JRFVoYanh_tZuF3VmeBVc5iNh8Avp4dE/edit?usp=sharing) podrá 
    encontrar un apoyo en hojas de cálculo para la identificación de requisitos funcionales y el progreso del proyecto.

    Autores: 
    * Juan Sebastián Londoño Ramírez [github](https://github.com/Juan3bitas)
    * Angie Carolina Grajales Giraldo [github](https://github.com/angiec-grajalesg)
    * Juan Diego García Nieto [github](https://github.com/Juan-Devgo)
  version: 1.1.1
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
...
  /api/lodgings:
    get:
      tags:
        - Lodging
      summary: Obtener todos los alojamientos
      description: Obtiene una lista de todos los alojamientos disponibles en el sistema.
      operationId: getLodgings
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
          description: Filtrar alojamientos por ciudad
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: float
          description: Filtrar alojamientos por precio mínimo
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: float
          description: Filtrar alojamientos por precio máximo
        - name: checkIn
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filtrar alojamientos por fecha de entrada (check-in) en formato AAAA-MM-DD HH:mm
        - name: checkOut
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filtrar alojamientos por fecha de salida (check-out) en formato AAAA-MM-DD HH:mm
      responses:
        '200':
          description: Alojamientos obtenidos satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamientos obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LodgingDTO'
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-10", message: "Error inesperado al tratar de obtener los alojamientos."}
    post:
      tags:
        - Lodging
      summary: Crear un nuevo alojamiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LodgingDTO'
      responses:
        '201':
          description: Alojamiento creado satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamiento creado satisfactoriamente."
                  content:
                    default: null                
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA10-1", message: "Error en la solicitud al crear un alojamiento."}
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA10-2", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA10-3", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '409':
          description: Conflicto - Alojamiento ya existe.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Conflicto - Alojamiento ya existe."
                  content:
                    default: {code: "RYA10-4", message: "El alojamiento ya se encuentra registrado."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-11", message: "Error inesperado al tratar de crear un alojamiento."}
      security:
        - bearerAuth: []
  /api/lodgings/{lodging_uuid}:
    get:
      tags:
        - Lodging
      summary: Obtener un alojamiento
      description: Obtiene un alojamiento específico identificado por su UUID.
      operationId: getLodging
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alojamiento obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamiento obtenido satisfactoriamente."
                  content:
                    $ref: '#/components/schemas/LodgingDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: { code: "getLodging401", message: "No autorizado para esta acción. (JWT ausente)" }
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: { code: "getLodging403", message: "Prohibido para esta acción (JWT inválido o sin permisos)" }
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "getLodging404", message: "No se encontró el alojamiento especificado para obtener."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-27", message: "Error inesperado al tratar de obtener un alojamiento."}
      security:
        - bearerAuth: []
    put:
      tags:
        - Lodging
      summary: Actualizar un alojamiento existente
      description: Actualiza los detalles de un alojamiento existente identificado por su UUID.
      operationId: updateLodging
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LodgingDTO'
      responses:
        '200':
          description: Alojamiento actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamiento actualizado satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA10-5", message: "Error en la solicitud al actualizar un alojamiento."}
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA10-6", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA10-7", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "RYA10-8", message: "No se encontró el alojamiento especificado para actualizar."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-12", message: "Error inesperado al tratar de actualizar el alojamiento."}
      security:
        - bearerAuth: []
    delete:
      tags:
        - Lodging
      summary: Eliminar un alojamiento existente (soft delete)
      description: Elimina un alojamiento existente identificado por su UUID de manera suave, cambiándole el estado a "DELETED".
      operationId: softDeleteLodging
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del alojamiento que se desea cambiarle el estado a "DELETED".
      responses:
        '200':
          description: Alojamiento cambiado de estado a "DELETED"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamiento eliminado satisfactoriamente."
                  content:
                    default: null
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA11-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA11-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "RYA11-3", message: "No se encontró el alojamiento especificado para eliminar."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-13", message: "Error inesperado al tratar de eliminar el alojamiento."}
      security:
        - bearerAuth: []
  /api/lodgings/{search}:
    get:
      tags:
        - Lodging
      summary: Buscar alojamientos por nombre o descripción
      description: Busca alojamientos cuyo nombre o descripción contengan el término de búsqueda proporcionado, cumpliendo con el requisito REQ-BQA-01 y REQ-BQA-04.
      operationId: searchLodgings
      parameters:
        - name: search
          in: path
          required: true
          schema:
            type: string
          description: Término de búsqueda para filtrar alojamientos por nombre o descripción.
      responses:
        '200':
          description: Alojamientos encontrados
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamientos obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LodgingDTO'
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-14", message: "Error inesperado al tratar de buscar los alojamientos."}
...
  /api/lodgings/{host_uuid}:
    get:
      tags:
        - Lodging
      summary: Obtener alojamientos de un anfitrión específico
      description: Obtiene una lista de alojamientos asociados a un anfitrión específico.
      operationId: getLodgingsByHostUUID
      parameters:
        - name: host_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del anfitrión para el cual se están obteniendo los alojamientos.
      responses:
        '200':
          description: Alojamientos obtenidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Alojamientos obtenidos satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/LodgingDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA09-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA09-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Anfitrión no encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Anifrión no encontrado."
                  content:
                    default: {code: "RYA09-3", message: "No se ha encontrado el anfitrión especificado para los alojamientos."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-16", message: "Error inesperado al tratar de obtener los alojamientos."}
      security:
        - bearerAuth: []
  /api/lodgings/{lodging_uuid}/comments:
    patch:
      tags:
        - Lodging
      summary: Agregar un comentario a un alojamiento
      description: Permite a un usuario agregar un comentario y una valoración a un alojamiento específico.
      operationId: addCommentToLodging
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del alojamiento al que se le va a agregar el comentario.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                valuation:
                  type: integer
                  format: float
              required:
                - comment
                - valuation
      responses:
        '201':
          description: Comentario agregado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Comentario agregado satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA05-1", message: "Error en la solicitud al tratar de agregar un comentario."}
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA05-2", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA05-3", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "RYA05-4", message: "No se encontró el alojamiento especificado para agregar el comentario."}
        '409':
          description: Conflicto - Comentario ya existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Conflicto - Comentario ya existe."
                  content:
                    default: {code: "RYA05-5", message: "Usted ya ha comentado en este alojamiento."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-17", message: "Error inesperado al tratar de agregar el comentario."}
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    LodgingDTO:
      description: Esquema que representa un alojamiento para transferencia de datos.
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        details:
          type: object
          properties:
            location:
              type: object
              properties:
                city:
                  type: string
                address:
                  type: string
                latitude:
                  type: number
                  format: float
                longitude:
                  type: number
                  format: float
            price:
              type: number
              format: float
            services:
              type: array
              items:
                type: string
            pictures:
              type: array
              items:
                type: string
                format: url
                example: "https://example.com/image.jpg"
      required:
        - title
        - description
        - details
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT